fix: Fix build error with VELOX_ENABLE_AGGREGATES=OFF
