+++ velox/functions/lib/sfm/CMakeLists.txt
@@ -12,14 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_executable(velox_functions_prestosql_aggregates_sfm_test SfmSketchTest.cpp)
+velox_add_library(velox_functions_sfm SfmSketch.cpp SfmSketchAccumulator.cpp)
 
-add_test(velox_functions_prestosql_aggregates_sfm_test
-         velox_functions_prestosql_aggregates_sfm_test)
+velox_link_libraries(velox_functions_sfm velox_common_base velox_memory
+                     Folly::folly)
 
-target_link_libraries(
-  velox_functions_prestosql_aggregates_sfm_test
-  velox_functions_prestosql_aggregates_sfm
-  velox_memory
-  GTest::gtest
-  GTest::gtest_main)
+if(${VELOX_BUILD_TESTING})
+  add_subdirectory(tests)
+endif()
