+++ velox/functions/prestosql/aggregates/CMakeLists.txt
@@ -69,10 +69,8 @@ velox_link_libraries(
   velox_functions_lib
   velox_functions_util
   velox_presto_types
-  velox_functions_prestosql_aggregates_sfm
+  velox_functions_sfm
   Folly::folly)
-
-add_subdirectory(sfm)
 
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(tests)
